requirements.txt
anyio==2.2.0
appdirs==1.4.4
arduino-python3==0.6
argon2-cffi==20.1.0
async-generator==1.10
atomicwrites==1.4.0
attrs==21.2.0
Babel==2.9.0
backcall==0.2.0
bleach==3.2.1
certifi==2020.11.8
cffi==1.14.3
chardet==4.0.0
colorama==0.4.4
control==0.8.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
docutils==0.16
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
h5py==3.2.1
idna==2.10
imageio==2.9.0
iniconfig==1.1.1
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyter-packaging==0.7.12
jupyter-server==1.4.1
jupyterlab==3.0.10
jupyterlab-pygments==0.1.2
jupyterlab-server==2.3.0
keyboard==0.13.5
Kivy==2.0.0
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.gstreamer==0.3.1
kivy-deps.sdl2==0.3.1
Kivy-examples==2.0.0
Kivy-Garden==0.1.4
kiwisolver==1.3.1
lcapy==0.70
llvmlite==0.36.0
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
mpmath==1.1.0
multithreading==0.1.9
mysql-connector-python==8.0.22
nbclassic==0.2.6
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.2
networkx==2.5
nodejs==0.1.1
notebook==6.1.4
numba==0.53.0
numpy==1.19.3
opencv-contrib-python==4.5.1.48
openpyxl==3.0.7
optional-django==0.1.0
packaging==20.9
pandas==1.1.4
pandoc==1.0.2
pandocfilters==1.4.3
parso==0.7.1
pdflatex==0.1.3
pickleshare==0.7.5
Pillow==8.0.1
pipenv==2020.11.4
plotly==4.14.3
pluggy==0.13.1
ply==3.11
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.13.0
py==1.10.0
PyBluez @ git+https://github.com/pybluez/pybluez@d7f36702bba2abc8ea41977b2eea56a17431ac6d
pycparser==2.20
pygame==2.0.0
Pygments==2.7.2
pynput==1.7.3
pyparsing==2.4.7
pypiwin32==223
PyQt5==5.15.1
PyQt5-sip==12.8.1
pyqtgraph==0.11.1
pyrsistent==0.17.3
pyrtlsdr==0.2.92
pyserial==3.4
pysigproc==0.4.0
pytest==6.2.4
python-dateutil==2.8.1
pytils==0.3
pytz==2020.4
PyWavelets==1.1.1
pywin32==228
pywinpty==0.5.7
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.25.1
retrying==1.3.3
scapy==2.4.4
scikit-commpy==0.6.0
scikit-image==0.18.1
scipy==1.6.1
seaborn==0.11.1
selection==0.0.14
Send2Trash==1.5.0
six==1.15.0
sniffio==1.2.0
socket.py==0.1.0
sympy==1.6.2
terminado==0.9.1
testpath==0.4.4
tifffile==2021.3.5
toml==0.10.2
tornado==6.1
traitlets==5.0.5
urllib3==1.26.2
user-agent==0.1.9
virtualenv==20.4.2
virtualenv-clone==0.5.4
wcwidth==0.2.5
webencodings==0.5.1
weblib==0.1.30
widgetsnbextension==3.5.1
xlrd==1.2.0
